Split up index.jsx more into component files.
