+++ src/MapInteraction.jsx
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
+import Controls from './Controls';
+
 import { clamp, distance, midpoint, touchPt, touchDistance } from './geometry';
 import makePassiveEventOption from './makePassiveEventOption';
 
@@ -402,136 +404,4 @@ class MapInteraction extends Component {
   }
 }
 
-/*
-  This component provides a map like interaction to any content that you place in it. It will let
-  the user zoom and pan the children by scaling and translating props.children using css.
-*/
-const MapInteractionCSS = (props) => {
-  return (
-    <MapInteraction {...props}>
-      {
-        ({ translation, scale }) => {
-          // Translate first and then scale.  Otherwise, the scale would affect the translation.
-          const transform = `translate(${translation.x}px, ${translation.y}px) scale(${scale})`;
-          return (
-            <div
-              style={{
-                height: '100%',
-                width: '100%',
-                position: 'relative', // for absolutely positioned children
-                overflow: 'hidden',
-                touchAction: 'none', // Not supported in Safari :(
-                msTouchAction: 'none',
-                cursor: 'all-scroll',
-                WebkitUserSelect: 'none',
-                MozUserSelect: 'none',
-                msUserSelect: 'none'
-              }}
-            >
-              <div
-                style={{
-                  transform: transform,
-                  transformOrigin: '0 0 '
-                }}
-              >
-                {props.children}
-              </div>
-            </div>
-          );
-        }
-      }
-    </MapInteraction>
-  );
-};
-
-class Controls extends Component {
-  render() {
-    const {
-      plusBtnContents,
-      minusBtnContents,
-      btnClass,
-      plusBtnClass,
-      minusBtnClass,
-      controlsClass,
-      scale,
-      minScale,
-      maxScale,
-      onClickPlus,
-      onClickMinus
-    } = this.props;
-
-    const btnStyle = { width: 30, paddingTop: 5, marginBottom: 5 };
-    const controlsStyle = controlsClass ? undefined : { position: 'absolute', right: 10, top: 10 };
-
-    function plusHandler(e) {
-      e.preventDefault();
-      e.target.blur();
-      onClickPlus();
-    }
-
-    function minusHandler(e) {
-      e.preventDefault();
-      e.target.blur();
-      onClickMinus();
-    }
-
-    return (
-      <div style={controlsStyle} className={controlsClass}>
-        <div>
-          <button
-            ref={(node) => { this.plusNode = node; }}
-            onClick={plusHandler}
-            onTouchEnd={plusHandler}
-            className={[
-              btnClass ? btnClass : '',
-              plusBtnClass ? plusBtnClass : '',
-            ].join(' ')}
-            type="button"
-            style={(btnClass || plusBtnClass) ? undefined : btnStyle}
-            disabled={scale >= maxScale}
-          >
-            {plusBtnContents}
-          </button>
-        </div>
-        <div>
-          <button
-            ref={(node) => { this.minusNode = node; }}
-            onClick={minusHandler}
-            onTouchEnd={minusHandler}
-            className={[
-              btnClass ? btnClass : '',
-              minusBtnClass ? minusBtnClass : '',
-            ].join(' ')}
-            type="button"
-            style={(btnClass || minusBtnClass) ? undefined : btnStyle}
-            disabled={scale <= minScale}
-          >
-            {minusBtnContents}
-          </button>
-        </div>
-      </div>
-    );
-  }
-}
-
-Controls.propTypes = {
-  onClickPlus: PropTypes.func.isRequired,
-  onClickMinus: PropTypes.func.isRequired,
-  plusBtnContents: PropTypes.node,
-  minusBtnContents: PropTypes.node,
-  btnClass: PropTypes.string,
-  plusBtnClass: PropTypes.string,
-  minusBtnClass: PropTypes.string,
-  controlsClass: PropTypes.string,
-  scale: PropTypes.number,
-  minScale: PropTypes.number,
-  maxScale: PropTypes.number
-};
-
-Controls.defaultProps = {
-  plusBtnContents: '+',
-  minusBtnContents: '-'
-};
-
-export { MapInteractionCSS, MapInteraction };
 export default MapInteraction;
