+++ src/index.js
@@ -1,0 +1,5 @@
+import MapInteractionCSS from './MapInteractionCSS';
+import MapInteraction from './MapInteraction';
+
+export { MapInteractionCSS, MapInteraction };
+export default MapInteraction;
